Add patch to disable 'erase_node' mco plugin
This patch just prevents erasing of MBR on each of node's disks and
rebooting node when user resets or deletes Fuel environment this node
belongs to. As Fuel is not able to provision POWER8 nodes this patch may
save time if it is needed to delete/reset environment with actually
'clean' POWER8 servers (e.g. no deployment tries, server only contains
nailgun-agent and mcollective).

Note, that if you want to delete actually deployed environment and reuse
POWER8 server for new one, if would be a better idea to reinstall Ubuntu
on it (as it will contain 'garbage' software from previous env), so in
such case you may not want to use this patch.

Change-Id: Iff742181c05dd4ea6c5b1d77bdcce863023a50f3